[ui] Use a hook to supply flattening/filtering on code locations
